+++ client/src/utils.rs
@@ -11,7 +11,7 @@ use crypto::{
     utils,
 };
 use iota_types::block::{
-    address::{Address, AliasAddress, Ed25519Address, NftAddress},
+    address::{Address, Ed25519Address},
     output::{AliasId, NftId},
     payload::TaggedDataPayload,
 };
@@ -35,16 +35,6 @@ pub fn bech32_to_hex(bech32: &str) -> Result<String> {
 pub fn hex_to_bech32(hex: &str, bech32_hrp: &str) -> Result<String> {
     let address: Ed25519Address = hex.parse::<Ed25519Address>()?;
     Ok(Address::Ed25519(address).to_bech32(bech32_hrp))
-}
-
-/// Transforms an alias id to a bech32 encoded address
-pub fn alias_id_to_bech32(alias_id: AliasId, bech32_hrp: &str) -> String {
-    Address::Alias(AliasAddress::new(alias_id)).to_bech32(bech32_hrp)
-}
-
-/// Transforms an nft id to a bech32 encoded address
-pub fn nft_id_to_bech32(nft_id: NftId, bech32_hrp: &str) -> String {
-    Address::Nft(NftAddress::new(nft_id)).to_bech32(bech32_hrp)
 }
 
 /// Transforms a prefix hex encoded public key to a bech32 encoded address
@@ -129,16 +119,16 @@ impl Client {
     /// Transforms an alias id to a bech32 encoded address
     pub async fn alias_id_to_bech32(&self, alias_id: AliasId, bech32_hrp: Option<&str>) -> crate::Result<String> {
         match bech32_hrp {
-            Some(hrp) => Ok(alias_id_to_bech32(alias_id, hrp)),
-            None => Ok(alias_id_to_bech32(alias_id, &self.get_bech32_hrp().await?)),
+            Some(hrp) => Ok(alias_id.to_bech32(hrp)),
+            None => Ok(alias_id.to_bech32(&self.get_bech32_hrp().await?)),
         }
     }
 
     /// Transforms an nft id to a bech32 encoded address
     pub async fn nft_id_to_bech32(&self, nft_id: NftId, bech32_hrp: Option<&str>) -> crate::Result<String> {
         match bech32_hrp {
-            Some(hrp) => Ok(nft_id_to_bech32(nft_id, hrp)),
-            None => Ok(nft_id_to_bech32(nft_id, &self.get_bech32_hrp().await?)),
+            Some(hrp) => Ok(nft_id.to_bech32(hrp)),
+            None => Ok(nft_id.to_bech32(&self.get_bech32_hrp().await?)),
         }
     }